refactor(debt): decompose settings-general god file (D-04)
Extract the ~500-LOC RemoteAccessSection into settings-remote-access.tsx as a
standalone component that re-acquires its contexts via hooks (no prop drilling,
reactivity preserved), and the shared SettingsRow into settings-row.tsx.
settings-general.tsx drops 1108 -> 589 LOC (under the 800 alert threshold).
app typecheck passes.

D-02: prompt-input.tsx is already decomposed into 12 prompt-input/ submodules
(5 tested); the remaining file is a coordinator — noted in the register.
